requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.2.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
multidict==4.5.2
six==1.12.0
typed-ast==1.4.0
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
